LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_setpointercapture_disconnected-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: DOMException InvalidStateError

Test Description: When the setPointerCapture method is invoked, if the target node does not participate in its ownerDocument's tree, a DOMException with the name InvalidStateError must be thrown.


Use the mouse, touch or pen to contact this box.
The following pointer types were detected: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN setPointerCapture: DOMException InvalidStateError
